graph backend: matches decompression and result mapping
Took some time to review and recall the continuation-passing-style to write the combinatorial tree traversal part.
See also:
- http://matt.might.net/articles/by-example-continuation-passing-style/
- http://okmij.org/ftp/continuations/

It would be wiser to move this match decompression as close as possible to the UI. (Thus, changing the data format between UI and graphd)  Also, if this is going to be done at the UI-side, we need to find a good alternative with minimal code modification that makes pausing the generator at arbitrary point possible since JavaScript lacks support for coroutine or call/cc.